notifications: Split out compose_notifications module.
