+++ tools/test-js-with-node
@@ -67,6 +67,7 @@ EXEMPT_FILES = make_set(
         "web/src/compose_call_ui.js",
         "web/src/compose_closed_ui.js",
         "web/src/compose_fade.js",
+        "web/src/compose_notifications.js",
         "web/src/compose_popovers.js",
         "web/src/compose_recipient.js",
         "web/src/compose_reply.js",
